refactor: 사용자 Service 인터페이스 삭제
